feat: move navigation logic to Dashboard component and remove from App component
